fix(compile): bypass FakeTensor bug that makes isinstance evaluation fail

+++ optimum/quanto/tensor/marlin/packed.py
@@ -158,7 +158,7 @@ class MarlinF8PackedTensor(torch.Tensor):
         assert data.dtype == torch.int32
         assert requires_grad is False
         return torch.Tensor._make_wrapper_subclass(
-            cls, size, strides=stride, dtype=torch.float8_e4m3fn, device=data.device, requires_grad=requires_grad
+            cls, size, strides=stride, dtype=torch.int32, device=data.device, requires_grad=requires_grad
         )
 
     def __init__(self, data, size, stride, requires_grad=False):
@@ -171,9 +171,6 @@ class MarlinF8PackedTensor(torch.Tensor):
     def pack(cls, tensor: torch.Tensor):
         out_features, in_features = tensor.shape
 
-        # TODO: getting
-        # AssertionError: Please convert all Tensors to FakeTensors first or instantiate FakeTensorMode with 'allow_non_fake_inputs'. Found in aten.bitwise_or_.Tensor(tensor([...], device='cuda:0', size=(512, 2048), dtype=torch.int32), FakeTensor(..., device='cuda:0', size=(512, 2048), dtype=torch.int32))
-        # here with torch.compile
         data_int32 = pack_fp8_as_int32(tensor.T)  # pack fp8 data to in32.
 
         perm = torch.empty(0, dtype=torch.int, device=tensor.device)
